requirements.txt
# Flask Framework
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3

# Flask Packages
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-User==1.0.2.2
flask_bcrypt==1.0.1

# MySQL
mysql-connector-python==8.0.32

paramiko==3.1.0